make ms per tick an integer

+++ src/grid.rs
@@ -266,7 +266,7 @@ impl canvas::Program<()> for Grid {
             frame.fill_text(Text {
                 content: format!(
                     "{} cell{} @ {} Ms/Tick, {:.3} Ticks/s.. Queued Ticks: {}",
-                    self.view.cells,
+                    self.view.cells as u64,
                     if self.view.cells == 1 { "" } else { "s" },
                     seconds_per_tick * 1000.0,
                     seconds_per_tick.recip(),
